Several adjustments were made
